requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.6.0
channels==4.2.0
channels-redis==4.2.1
celery==5.4.0
redis==5.2.1
pillow==11.0.0
stripe==10.1.0
python-decouple==3.8
geopy==2.4.1